move android PA functionality to scenarios folder
