+++ samples/java/android/sdkdemo/app/src/main/java/com/microsoft/cognitiveservices/speech/samples/sdkdemo/MainActivity.java
@@ -86,8 +86,6 @@ public class MainActivity extends AppCompatActivity {
     private Button recognizeContinuousButton;
     private Button recognizeIntentButton;
     private Button recognizeWithKeywordButton;
-    private Button pronunciationAssessmentButton;
-    private Button pronunciationAssessmentFromStreamButton;
 
     private MicrophoneStream microphoneStream;
     private MicrophoneStream createMicrophoneStream() {
@@ -116,8 +114,6 @@ public class MainActivity extends AppCompatActivity {
         recognizeContinuousButton = findViewById(R.id.buttonRecognizeContinuous);
         recognizeIntentButton = findViewById(R.id.buttonRecognizeIntent);
         recognizeWithKeywordButton = findViewById(R.id.buttonRecognizeWithKeyword);
-        pronunciationAssessmentButton = findViewById(R.id.buttonPronunciationAssessment);
-        pronunciationAssessmentFromStreamButton = findViewById(R.id.buttonPronunciationAssessmentStream);
 
         // Initialize SpeechSDK and request required permissions.
         try {
@@ -419,190 +415,6 @@ public class MainActivity extends AppCompatActivity {
                 }
             }
         });
-
-
-        ///////////////////////////////////////////////////
-        // pronunciation assessment from file
-        ///////////////////////////////////////////////////
-        pronunciationAssessmentButton.setOnClickListener(view ->  {
-            final String logTag = "pron";
-
-            disableButtons();
-            clearTextBox();
-
-            try {
-                final AudioConfig audioConfig = AudioConfig.fromWavFileInput(copyAssetToCacheAndGetFilePath("pronunciation_assessment.wav"));
-                // Switch to other languages for example Spanish, change language "en-US" to "es-ES". Language name is not case sensitive.
-                final SpeechRecognizer reco = new SpeechRecognizer(speechConfig, "en-US", audioConfig);
-
-                String referenceText =  "Hello world hello. Today is a nice day.";
-                PronunciationAssessmentConfig pronConfig =
-                    new PronunciationAssessmentConfig(referenceText,
-                        PronunciationAssessmentGradingSystem.HundredMark,
-                        PronunciationAssessmentGranularity.Phoneme);
-
-                pronConfig.applyTo(reco);
-
-                ArrayList<String> recognizedWords = new ArrayList<>();
-                ArrayList<Word> pronWords = new ArrayList<>();
-                long[] totalDurations = new long[2]; // duration, valid duration
-                long[] offsets = new long[2]; // start offset, end offset
-                double[] totalAccuracyScore = new double[1];
-                reco.recognized.addEventListener((o, speechRecognitionResultEventArgs) -> {
-                    final String s = speechRecognitionResultEventArgs.getResult().getText();
-                    Log.i(logTag, "Final result received: " + s);
-                    PronunciationAssessmentResult pronResult = PronunciationAssessmentResult.fromResult(speechRecognitionResultEventArgs.getResult());
-                    Log.i(logTag, "Accuracy score: " + pronResult.getAccuracyScore() +
-                        ";  pronunciation score: " +  pronResult.getPronunciationScore() +
-                        ", completeness score: " + pronResult.getCompletenessScore() +
-                        ", fluency score: " + pronResult.getFluencyScore());
-                    String jString = speechRecognitionResultEventArgs.getResult().getProperties().getProperty(PropertyId.SpeechServiceResponse_JsonResult);
-
-                    for (WordLevelTimingResult w: pronResult.getWords()) {
-                        Word word = new Word(w.getWord(),
-                            w.getErrorType(),
-                            w.getAccuracyScore(),
-                            w.getDuration() / 10000, // The time unit is ticks (100ns); convert it to milliseconds
-                            w.getOffset() / 10000); // Same as above
-                        pronWords.add(word);
-                        recognizedWords.add(word.word);
-                        totalAccuracyScore[0] += word.duration * word.accuracyScore;
-                        totalDurations[0] += word.duration;
-                        if (word.errorType != null && word.errorType.equals("None")) {
-                            totalDurations[1] += word.duration + 10;
-                        }
-                        offsets[1] = word.offset + word.duration + 10;
-                    }
-                });
-
-                reco.sessionStopped.addEventListener((o, s) -> {
-                    Log.i(logTag, "Session stopped.");
-                    reco.stopContinuousRecognitionAsync();
-
-                    // We can calculate whole accuracy score by duration weighted averaging
-                    double accuracyScore = totalAccuracyScore[0] / totalDurations[0];
-                    double fluencyScore = 0;
-                    if (!recognizedWords.isEmpty()) {
-                        offsets[0] = pronWords.get(0).offset;
-                        fluencyScore = (double)totalDurations[1] / (offsets[1] - offsets[0]) * 100;
-                    }
-
-                    // For continuous pronunciation assessment mode, the service won't return the words with `Insertion` or `Omission`
-                    // even if miscue is enabled.
-                    // We need to compare with the reference text after received all recognized words to get these error words.
-                    String[] referenceWords = referenceText.toLowerCase().split(" ");
-                    for (int j = 0; j < referenceWords.length; j++) {
-                        referenceWords[j] = referenceWords[j].replaceAll("^\\p{Punct}+|\\p{Punct}+$","");
-                    }
-                    Patch<String> diff = DiffUtils.diff(Arrays.asList(referenceWords), recognizedWords, true);
-
-                    int currentIdx = 0;
-                    ArrayList<Word> finalWords = new ArrayList<>();
-                    int[] validWord = new int[1];
-                    for (AbstractDelta<String> d : diff.getDeltas()) {
-                        if (d.getType() == DeltaType.EQUAL) {
-                            for (int i = currentIdx; i < currentIdx + d.getSource().size(); i++) {
-                                finalWords.add(pronWords.get(i));
-                            }
-                            currentIdx += d.getTarget().size();
-                            validWord[0] += d.getTarget().size();
-                        }
-                        if (d.getType() == DeltaType.DELETE || d.getType() == DeltaType.CHANGE) {
-                            for (String w : d.getSource().getLines()) {
-                                finalWords.add(new Word(w, "Omission"));
-                            }
-                        }
-                        if (d.getType() == DeltaType.INSERT || d.getType() == DeltaType.CHANGE) {
-                            for (int i = currentIdx; i < currentIdx + d.getTarget().size(); i++) {
-                                Word w = pronWords.get(i);
-                                w.errorType = "Insertion";
-                                finalWords.add(w);
-                            }
-                            currentIdx += d.getTarget().size();
-                        }
-                    }
-
-                    // Calculate whole completeness score
-                    double completenessScore = (double)validWord[0] / referenceWords.length * 100;
-
-                    AppendTextLine("Paragraph accuracy score: " + accuracyScore +
-                        ", completeness score: " +completenessScore +
-                        " , fluency score: " + fluencyScore + "\n", true);
-                    for (Word w : finalWords) {
-                        AppendTextLine(" word: " + w.word + "\taccuracy score: " +
-                            w.accuracyScore + "\terror type: " + w.errorType, false);
-                    }
-
-                    enableButtons();
-                });
-
-                reco.startContinuousRecognitionAsync();
-            } catch (Exception ex) {
-                System.out.println(ex.getMessage());
-                displayException(ex);
-            }
-        });
-
-        ///////////////////////////////////////////////////
-        // pronunciation assessment from stream
-        ///////////////////////////////////////////////////
-        pronunciationAssessmentFromStreamButton.setOnClickListener(view -> {
-            final String logTag = "pron";
-
-            if (this.microphoneStream != null) {
-                /*
-                 The pronunciation assessment service has a longer default end silence timeout (5 seconds) than normal STT
-                 as the pronunciation assessment is widely used in education scenario where kids have longer break in reading.
-                 You can explicitly stop the recording stream to stop assessment and the service will return the results immediately
-                 */
-                this.releaseMicrophoneStream();
-                return;
-            }
-
-            disableButtons();
-            this.pronunciationAssessmentFromStreamButton.setEnabled(true);
-            this.pronunciationAssessmentFromStreamButton.setText("Stop recording");
-            clearTextBox();
-            try {
-                 createMicrophoneStream();
-                final AudioConfig audioConfig = AudioConfig.fromStreamInput(createMicrophoneStream());
-                // Switch to other languages for example Spanish, change language "en-US" to "es-ES". Language name is not case sensitive.
-                final SpeechRecognizer reco = new SpeechRecognizer(speechConfig, "en-US", audioConfig);
-
-                // Replace this referenceText to match your input.
-                String referenceText =  "Today is a nice day.";
-                PronunciationAssessmentConfig pronConfig =
-                    new PronunciationAssessmentConfig(referenceText,
-                        PronunciationAssessmentGradingSystem.HundredMark,
-                        PronunciationAssessmentGranularity.Phoneme);
-
-                pronConfig.applyTo(reco);
-
-                reco.recognized.addEventListener((o, speechRecognitionResultEventArgs) -> {
-                    final String s = speechRecognitionResultEventArgs.getResult().getText();
-                    AppendTextLine("Final result received: " + s, true);
-                    PronunciationAssessmentResult pronResult = PronunciationAssessmentResult.fromResult(speechRecognitionResultEventArgs.getResult());
-                    AppendTextLine("Accuracy score: " + pronResult.getAccuracyScore() +
-                        ";  pronunciation score: " +  pronResult.getPronunciationScore() +
-                        ", completeness score: " + pronResult.getCompletenessScore() +
-                        ", fluency score: " + pronResult.getFluencyScore(), false);
-                });
-
-                reco.sessionStopped.addEventListener((o, s) -> {
-                    Log.i(logTag, "Session stopped.");
-                    reco.stopContinuousRecognitionAsync();
-
-                    this.releaseMicrophoneStream();
-                    enableButtons();
-                    this.pronunciationAssessmentFromStreamButton.setText("Pronunciation Assessment From Stream");
-                });
-
-                reco.recognizeOnceAsync();
-            } catch (Exception ex) {
-                System.out.println(ex.getMessage());
-                displayException(ex);
-            }
-        });
     }
 
 
@@ -636,8 +448,6 @@ public class MainActivity extends AppCompatActivity {
             recognizeContinuousButton.setEnabled(false);
             recognizeIntentButton.setEnabled(false);
             recognizeWithKeywordButton.setEnabled(false);
-            pronunciationAssessmentButton.setEnabled(false);
-            pronunciationAssessmentFromStreamButton.setEnabled(false);
         });
     }
 
@@ -648,8 +458,6 @@ public class MainActivity extends AppCompatActivity {
             recognizeContinuousButton.setEnabled(true);
             recognizeIntentButton.setEnabled(true);
             recognizeWithKeywordButton.setEnabled(true);
-            pronunciationAssessmentButton.setEnabled(true);
-            pronunciationAssessmentFromStreamButton.setEnabled(true);
         });
     }
 
@@ -688,24 +496,5 @@ public class MainActivity extends AppCompatActivity {
     private static ExecutorService s_executorService;
     static {
         s_executorService = Executors.newCachedThreadPool();
-    }
-
-    public class Word {
-        public String word;
-        public String errorType;
-        public double accuracyScore;
-        public long duration;
-        public long offset;
-        public Word(String word, String errorType) {
-            this.word = word;
-            this.errorType = errorType;
-        }
-
-        public Word(String word, String errorType, double accuracyScore, long duration, long offset) {
-            this(word, errorType);
-            this.accuracyScore = accuracyScore;
-            this.duration = duration;
-            this.offset = offset;
-        }
     }
 }
