+++ samples/java/android/sdkdemo/app/src/main/res/layout/activity_main.xml
@@ -7,22 +7,6 @@
     tools:context="com.microsoft.cognitiveservices.speech.samples.sdkdemo.MainActivity">
 
   <Button
-      android:id="@+id/buttonPronunciationAssessmentStream"
-      android:layout_width="336dp"
-      android:layout_height="wrap_content"
-      android:layout_marginStart="8dp"
-      android:layout_marginEnd="8dp"
-      android:layout_marginBottom="8dp"
-      android:text="Pronunciation Assessment From Stream"
-      android:textAllCaps="false"
-      app:layout_constraintBottom_toTopOf="@+id/recognizedText"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintHorizontal_bias="0.406"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@id/buttonPronunciationAssessment"
-      app:layout_constraintVertical_bias="1.0" />
-
-  <Button
       android:id="@+id/buttonRecognize"
       android:layout_width="336dp"
       android:layout_height="wrap_content"
@@ -33,7 +17,9 @@
       app:layout_constraintBottom_toTopOf="@+id/buttonRecognizeIntermediate"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent" />
+      app:layout_constraintTop_toTopOf="parent"
+      app:layout_constraintVertical_chainStyle="packed"
+      app:layout_constraintVertical_bias="0" />
 
   <Button
       android:id="@+id/buttonRecognizeIntermediate"
@@ -77,20 +63,6 @@
       app:layout_constraintTop_toBottomOf="@+id/buttonRecognizeContinuous" />
 
   <Button
-      android:id="@+id/buttonPronunciationAssessment"
-      android:layout_width="336dp"
-      android:layout_height="wrap_content"
-      android:layout_marginStart="8dp"
-      android:layout_marginEnd="8dp"
-      android:layout_marginBottom="8dp"
-      android:text="Pronunciation Assessment"
-      android:textAllCaps="false"
-      app:layout_constraintBottom_toTopOf="@id/buttonPronunciationAssessmentStream"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@+id/buttonRecognizeWithKeyword" />
-
-  <Button
       android:id="@+id/buttonRecognizeWithKeyword"
       android:layout_width="336dp"
       android:layout_height="wrap_content"
@@ -99,7 +71,7 @@
       android:layout_marginBottom="8dp"
       android:text="Recognize with keyword"
       android:textAllCaps="false"
-      app:layout_constraintBottom_toTopOf="@+id/buttonPronunciationAssessment"
+      app:layout_constraintBottom_toTopOf="@+id/recognizedText"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintHorizontal_bias="0.406"
       app:layout_constraintStart_toStartOf="parent"
@@ -119,6 +91,6 @@
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintHorizontal_bias="0.0"
       app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@+id/buttonPronunciationAssessment" />
+      app:layout_constraintTop_toBottomOf="@+id/buttonRecognizeWithKeyword" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
