+++ scenarios/java/android/language-learning/app/.gitignore
@@ -1,0 +1,1 @@
+/build
