+++ scenarios/java/android/language-learning/app/src/main/assets/pronunciation_assessment.wav
