+++ scenarios/java/android/language-learning/app/src/main/java/com/microsoft/cognitiveservices/speech/samples/sdkdemo/MicrophoneStream.java
@@ -1,0 +1,63 @@
+//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE.md file in the project root for full license information.
+//
+package com.microsoft.cognitiveservices.speech.samples.sdkdemo;
+
+import android.media.AudioFormat;
+import android.media.AudioRecord;
+import android.media.MediaRecorder;
+import android.util.Log;
+
+import com.microsoft.cognitiveservices.speech.audio.PullAudioInputStreamCallback;
+import com.microsoft.cognitiveservices.speech.audio.AudioStreamFormat;
+
+/**
+ * MicrophoneStream exposes the Android Microphone as an PullAudioInputStreamCallback
+ * to be consumed by the Speech SDK.
+ * It configures the microphone with 16 kHz sample rate, 16 bit samples, mono (single-channel).
+ */
+public class MicrophoneStream extends PullAudioInputStreamCallback {
+    private final static int SAMPLE_RATE = 16000;
+    private final AudioStreamFormat format;
+    private AudioRecord recorder;
+
+    public MicrophoneStream() {
+        this.format = AudioStreamFormat.getWaveFormatPCM(SAMPLE_RATE, (short)16, (short)1);
+        this.initMic();
+    }
+
+    public AudioStreamFormat getFormat() {
+        return this.format;
+    }
+
+    @Override
+    public int read(byte[] bytes) {
+        if (this.recorder != null) {
+            long ret = this.recorder.read(bytes, 0, bytes.length);
+            return (int) ret;
+        }
+        return 0;
+    }
+
+    @Override
+    public void close() {
+        this.recorder.release();
+        this.recorder = null;
+    }
+
+    private void initMic() {
+        // Note: currently, the Speech SDK support 16 kHz sample rate, 16 bit samples, mono (single-channel) only.
+        AudioFormat af = new AudioFormat.Builder()
+                .setSampleRate(SAMPLE_RATE)
+                .setEncoding(AudioFormat.ENCODING_PCM_16BIT)
+                .setChannelMask(AudioFormat.CHANNEL_IN_MONO)
+                .build();
+        this.recorder = new AudioRecord.Builder()
+                .setAudioSource(MediaRecorder.AudioSource.VOICE_RECOGNITION)
+                .setAudioFormat(af)
+                .build();
+
+        this.recorder.startRecording();
+    }
+}
