+++ scenarios/java/android/language-learning/app/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_name">SpeechSDKQuickstart</string>
+    <string name="speechButtonText">Recognize Speech</string>
+</resources>
