+++ scenarios/java/android/language-learning/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = "Speech Recognition"
+include ':app'
